chore: move `CodeEditorColors` to constants file
