+++ app/client/src/globalStyles/CodemirrorHintStyles.ts
@@ -1,8 +1,10 @@
 import { createGlobalStyle } from "styled-components";
 import type { EditorTheme } from "components/editorComponents/CodeEditor/EditorConfig";
 import type { Theme } from "constants/DefaultTheme";
-import { LINT_TOOLTIP_JUSTIFIED_LEFT_CLASS } from "components/editorComponents/CodeEditor/constants";
-import { CodeEditorColors } from "components/editorComponents/CodeEditor/styledComponents";
+import {
+  CodeEditorColors,
+  LINT_TOOLTIP_JUSTIFIED_LEFT_CLASS,
+} from "components/editorComponents/CodeEditor/constants";
 
 export const CodemirrorHintStyles = createGlobalStyle<{
   editorTheme: EditorTheme;
