+++ app/client/src/utils/autocomplete/ternDocTooltip.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import { ternDocsInfo } from "@appsmith/utils/autocomplete/EntityDefinitions";
 import type { Completion, TernCompletionResult } from "./CodemirrorTernService";
-import { CodeEditorColors } from "components/editorComponents/CodeEditor/styledComponents";
+import { CodeEditorColors } from "components/editorComponents/CodeEditor/constants";
 import { Link } from "design-system";
 
 export function renderTernTooltipContent(
